New #DISTANCE directives: NOTBONDED [atoms or types]                           NOTPIVOT [atoms or types]
Specifically excludes the specified atoms from being either
pivots or bonded atoms. Useful because: if I want all the bonds
from peaks to non-peaks, I can now say PIVOT Q / NOTBONDED Q and
it will do the trick.

New #DISTANCE option: OUTPUT LIST=DEBUG

Every time you issue INCLUDE/PIVOT/BONDED/EXCLUDE/NOTBONDED/NOTPIVOT this
option prints a list of all the atoms along with the three column vectors
which specify where an atom can appear in a distance or angle result.
Useful to see what exactly is going on.

New #DISTANCE option: OUTPUT MONITOR=PEAKSTOGUI

For each pivot atom this option outputs the pivot with its SPARE value,
and the distance to the closest of the bonded atoms and its name -
(Like this: ^^WI 'Q(1) height: 1.43, 1.243A from C(11)') - allowing
it to be added to a listbox in the GUI. You probably want to use the options
PIVOT Q and NOTBONDED Q in conjunction with this option. See script XPRUNE
for an example.