Right - this is a useful one and I'm quite pleased with it.
Added a new lexical token PART to the scanner. The following
examples show what you might do with it:
----
Get distances between all atoms in part 1:
  #DIST
  INCLUDE PART(1)
  END
----
Competitively refine the occupancy of two parts, e.g. in disorder:
  #LIST 12
  FULL X'S U'S
  EQUIV PART(1,OCC) PART(2,OCC)
  WEIGHT -1 PART(1,OCC)
  END
----
Quickly id whole fragments in regularise:
  #REGUL
  target c(1) to c(6)
  ideal c(11) to c(16)
  cameron
  map part(2)
  onto part(1)
  END
----
Move a part of the structure:
  #EDIT
  ADD 0.2 PART(3,X)
  END
----
Have different regimes for refining e.g. a host and a guest molecule:
  #LIST 12
  FULL PART(0,X'S,U'S) PART(1,X'S,U[ISO])
  END
----
So it's quite handy. Part numbers are stored in the 15th slot of LIST 5, and
are defined as follows:
  ( 1000 * GROUPNUMBER ) + COMPONENTNUMBER.
    e.g. 123456 is group 123 and component 456.
  The same syntax is used to select a part in the lexical input:
    e.g. PART(123456) selects atoms in group 123, component 456.
  BUT, if the value of the group or component is 999, all atoms in that
  group or component are selected.
    e.g. PART(999003) selects all atoms with component 3 in their partnumber and
    PART(4999) selects all atoms in group 4.

The reason for having groups and parts is to enable logical grouping of, for example
a disorder where all affected atoms are group 1, with each component being part 0,1,2 etc.

NB. #script xparts runs a useful tool for defining part numbers.