Remove debugging from PART processing.
For every error message that prints a column number, output a star below the
offending line in the offending column.

In KLXRDD, return with an error on finding an invalid operator. Previously
it muddled on, but all the card pointers were set negative causing trouble
later.

Remove all NCAWU output as part of an ongoing battle.

In KATOMU, don't allow UNTIL to or from a PART - KATOMU gets called
repeatedly for a single PART and would therefore return a range for every
atom in the PART.