In function MAKE41 (constructs a list of bonds from atom coords), add extra logic so that any symmetry operation that is not I, and does not move an atom is ignored. This prevents two bonds being formed to atoms on special positions. This mattered not for display, but it matters much to Mogul.
The BREAK directive in L40 breaks bonds that would otherwise form and be
stored in L41. If you explicitly give a symmetry operator,S, of zero,
then all bonds between the pair of atoms will be broken, even if related
by symmetry. For example: Imagine a Zn(1) on a 2-fold axis. Imagine it
is coordinated by a carboxylate anion. Imagine that the C(1) atom is
accidentally bonding to the Zn(1) - it shouldn't only the O's should
bond. You say "BREAK ZN(1) TO C(1)", but the ZN(1) TO C(1,2) remains.
Now you can say "BREAK ZN(1,0) TO C(1,0)" and all ZN(1)-C(1) bonds are
broken, hurrah.