Increase maximum parts per parameter (MXPPP) from 10 to 50.
In KDIST4 (analagous to KDIST1, but returns distances from
List 41) keep track of the List 12 pointers if required.

Made XCOVAR more consistent - the VCV returned for distances
or angles is now always pre and postmultiplied by the
matrix of constraint for the given atoms. The 'weight' vector
that is returned is now always a unit matrix. Previously
XCOVAR treated atoms with single part and multi part
paramters differently.

Also, the diagonal multiplier matrix that was computed for
atoms with multipart parameters (XFWM) isn't required since
in these cases the multiplier information is already applied
to the VCV: a unit matrix is now returned.