Double size of store arrays to 64Mb.
If a directive has a maximum count of >= 9999, then treat this as a
special case, allowing an infinite number of directives of this
type to be issued. (Suggest that if any directives really do
need a maximum count to prevent, e.g. a runaway loop somewhere, then
they be set to 9998).